fix all mistaces

+++ src/main/java/ru/pipDota2/web/ArticleController.java
@@ -1,6 +1,7 @@
 package ru.pipDota2.web;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -8,6 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import ru.pipDota2.domain.Article;
+import ru.pipDota2.domain.Role;
 import ru.pipDota2.service.ArticleServiceImpl;
 import ru.pipDota2.service.UserService;
 import ru.pipDota2.web.forms.Error;
@@ -40,11 +42,15 @@ public class ArticleController {
     }
 
     @GetMapping("/delete/article")
+    @PreAuthorize("hasRole('ADMIN')")
     public Result deleteArticle(@RequestParam("article_id") int articleId){
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         try {
             org.springframework.security.core.userdetails.User user =
                     (org.springframework.security.core.userdetails.User)authentication.getPrincipal();
+            if (!user.getAuthorities().contains(Role.ADMIN)){
+                return new Error("Недостаточно прав");
+            }
 
             String username = user.getUsername();
             if (articleService.deleteArticleByUser(articleId,
@@ -61,14 +67,18 @@ public class ArticleController {
     }
 
     @GetMapping("/save/article")
+    @PreAuthorize("hasRole('ADMIN')")
     public Result SaveArticle(@RequestParam("title")String title,
                                @RequestParam("content")String content){
         try {
             Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
             org.springframework.security.core.userdetails.User user =
                     (org.springframework.security.core.userdetails.User)authentication.getPrincipal();
-            String username = user.getUsername();
+            if (!user.getAuthorities().contains(Role.ADMIN)){
+                return new Error("Недостаточно прав");
+            }
 
+            String username = user.getUsername();
             if(articleService.saveArticle(title, content,
                     userService.findUserByName(username).orElseThrow(() ->
                             new UsernameNotFoundException("user " + username + " was not found")))){
